requirements.txt
pandas
PyYAML
requests
plotly
pymongo
SQLAlchemy
